[Hnswlib] Removing redundunt code
Hnswlib is using methods like `(L2Sqr|InnerProduct)SIMD16Ext[F16]`
that iterate by 16 components, while the maximum parallelism could be
8 with AVX instructions and 4 with SSE. Existing code was simply inlining
2x8 AVX instructions or 4x4 SSE to compuate x16 componets per iterations.
Presumably this was done to achive some optimizations.

In this commit we remove duplicated blocks in distance computation and
switch to iterations of 8 `(L2Sqr|InnerProduct)SIMD8Ext[F16]`.

While this didn't introduce any measurable regression in performance for
embeddings of size multiple 16 it allows to unblock faster vectorized
computation for dimensions 8 (AVX instead of SSE) and dimensions that were
multiples of 8, but not of 16. It also brings improvement in computing
risiduals computation for the same reasons.

Change-Id: I12f702a693d4cef480b958bb62d3acbf6b08d66f
JIRA: RECO-1789