[Bruteforce] Adding inplace bruteforce search for HNSW indices
To compute HNSW Recall of each index we need to perform bruteforce distance
computation on all the items avaialble in the index. Currently this is
possible by creating bruteforce index and populating it with vectors from
HNSW index. This doubles memory footprint and creates additional overhead
related to copy.

To solve this and improve offline jobs robustness following changes were
made:
  1. Bruteforce "algorithm" was factored out into reusable class
  2. From HnswIndex.search method `preprocessItem` method was extracted to
  allow reusing preprocessing logic (normalization, precision).
  3. `searchBruteforce` method was added to allow calling in-place bruteforce
  algorithm on the existing HnswIndex.

Change-Id: I89174cc30e4491221f4d528be38fe3ba1fd58c3a